Handle visbility map updates in WAL redo

+++ pageserver/src/walredo.rs
@@ -226,7 +226,7 @@ impl WalRedoManagerInternal {
         let start = Instant::now();
 
         let apply_result: Result<Bytes, Error>;
-        if tag.rel.forknum == pg_constants::PG_XACT_FORKNUM as u8 {
+        if tag.rel.forknum == pg_constants::PG_XACT_FORKNUM {
             const ZERO_PAGE: [u8; 8192] = [0u8; 8192];
             let mut page = BytesMut::new();
             if let Some(fpi) = base_img {
